refactor: move User and Role entities to the identity domain package
