+++ backend/src/main/java/com/ecommerce/identity/domain/Role.java
@@ -1,4 +1,4 @@
-package com.ecommerce.model;
+package com.ecommerce.identity.domain;
 
 public enum Role {
     ADMIN,
