+++ backend/src/test/java/com/ecommerce/service/OrderServiceTest.java
@@ -24,11 +24,11 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.ecommerce.dto.OrderDTO;
 import com.ecommerce.dto.OrderItemDTO;
 import com.ecommerce.dto.OrderResponseDTO;
+import com.ecommerce.identity.domain.User;
 import com.ecommerce.model.Order;
 import com.ecommerce.model.OrderItem;
 import com.ecommerce.model.OrderStatus;
 import com.ecommerce.model.Product;
-import com.ecommerce.model.User;
 import com.ecommerce.repository.OrderRepository;
 import com.ecommerce.repository.ProductRepository;
 import com.ecommerce.repository.UserRepository;
